Create PDFMerger & Watermark add-on.py
It's a simple code to use. Whoever wants to merge their PDFs simply run this code and provide answers to each question. It will automatically show merging files and merged file names. There is an add-on question where the user can attach a watermark on each page of the combined PDF file. I welcome all the feedback and looking forward to improving the script in the future.